requirements.txt
# Core
pandas==2.2.2
numpy==1.26.4
sqlalchemy==2.0.30
pymysql==1.1.1          # for MySQL
psycopg2-binary==2.9.9  # for PostgreSQL

# Machine Learning
scikit-learn==1.5.0
xgboost==2.0.3
imbalanced-learn==0.12.3
shap==0.45.1

# Visualization (optional local plots)
matplotlib==3.9.0
seaborn==0.13.2

# Scheduling
schedule==1.2.2

# Utilities
python-dotenv==1.0.1
loguru==0.7.2
requests==2.32.3
openpyxl==3.1.3
joblib==1.4.2
